improve profile-page detection, and more
The main change here is the detection of a user's profile page, which is something we can distill (unlike other pages on X which may have a similar path-style/pattern).

Additionally, this change introduces a small amount of refactoring, as well as a bug fix (need to invoke `isSupportedPage`).

Lastly, this change introduces a small preface to the in-situ metadata, to assist the LLM in understanding where the actual page content begins within the distillation result.