[cr133] RenderCoordinator methods are exposed as functions.
Chromium change:

https://source.chromium.org/chromium/_/chromium/devtools/devtools-frontend/+/328b3a7c2d447b988e7e26e87af1d6ba30848894

[cleanup] Expose `RenderCoordinator` methods as functions.

This is step one to making the `RenderCoordinator` more ergonomic to use
(and our codebase slightly more readable): Instead of going via the
singleton all the time, just expose the `RenderCoordinator` methods as
functions on the `RenderCoordinator` module. That makes code shorter and
easier on the eyes, and also doesn't leave the engineer wondering why on
earth the `RenderCoordinator` is a class.

The second step will be to replace the implementation of the render
coordinator with just functions instead of the singleton.

Bug: 387186170