[Android] Temporarily disable ChromeOS sync tests
A new settings srceen is used for ChromeOS now, so these tests need adjustments.
Since it is not a quick fix, it will be done in a separate issue https://github.com/brave/brave-browser/issues/43285

Chromium change:
https://chromium.googlesource.com/chromium/src/+/143e5dd239772d488aa172cb7c712cbde6d53f93

[Android][Signin] Remove deprecated promo in settings

The sign-in promo in settings is deprecated with the UNO2 launch. This
CL removes the related code and flag to clean-up the logic.

Bug: 350461111