Simplify broken code
Since 2c47a166ab46be3afe75db803f7994ce1dce4856, `npm start` used to be
able to run Brave's codesigned .app. The recent changes to
build/mac/BUILD.GN no longer produce this .app. The reason for this is
that upstream's signing code, which we now use, creates the signed .app
in a temporary directory, which then gets packaged into the .pkg and
.dmg artifacts. It does not seem worth the effort to restore this
behavior. If a dev wants to run the codesigned .app, then they can
extract it from the .dmg.