[Android] Changes for partner bookmark initialization
Chromium change:
https://chromium.googlesource.com/chromium/src/+/25845008b79c77a30986f99ee9bcbff4a81d10f6

Fix partner bookmark initialization.

PartnerBookmarks were previously being initialized in
ChromeActivity, but not all usage of the BookmarkBridge comes
from ChromeActivity derivatives. In particular if you enable
Don't Keep Activities in Android and load the BookmarkActivity,
then leave, then refocus from recents, it would crash on an
assertion that the partner bookmarks weren't loaded.

To fix the above issue, this moves the PartnerBookmark init
into the process initialization code path that runs once per
application init.

Subsequently, I was trying to move around when the initialization
happened and the internal implementation crashes because it can
not be called on the main thread (no idea why we aren't seeing
this in the wild), so this adds a async way of fetching the down
stream implementation (subsequent patch will utilize it).

Bug: 384197258