[Android] Pass profile reference to BookmarkActivities
Chromium change:
https://chromium.googlesource.com/chromium/src/+/e1fc3080db043124e75031b233f0808c057cc78c

Pass profile reference to BookmarkActivities.

Instead of loading getLastUsedRegularProfile, the calling activity will
serialize the current Profile state into the activity and the receiving
BookmarkActivity will use that to load the correct Profile.

While this CL enables the profile loading capability into the base
class of SynchronousInitializationActivity, it is only used by
BookmarkActivity and all other activities should use the legacy code path.

This will be the first production usage of ProfileResolver.

One assertion made in SynchronousInitializationActivity is that if the
profile is not resolved, then the activity will be directly finished.
This will happen in cases where the activity was started with an
incognito profile, then the task is killed by the OS (for debugging,
turn on don't keep activity in Android's developer settings), and
then restarted from recents. I don't think any of these activities
should be recreating the incognito profile that has since been destroyed, and instead should just return to the activity below it in
the task. In the future, we could consider variable recovery paths, but
I think that would be premature optimizations at this point.

Bug: 40254448