[cr135] `GetSearchProvidersUsingKeywordResult` with resolver arg
This change adjusts `GetSearchProvidersUsingKeywordResult` to include
the resolver argument to the override of this upstream function. This
affects other caller sites for this function, including a unit test,
which required the argument too. In order to correct the test, the
previous members were replaced with an instance of
`SearchEnginesTestEnvironment`, which does a similar job, but in a much
simpler way, whislt providing also the necessary resolver instance.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/d8bd0119166870ca42ab64af69487cecbe8298db

commit d8bd0119166870ca42ab64af69487cecbe8298db
Author: Nicolas Dossou-Gbete <dgn@chromium.org>
Date:   Thu Feb 13 00:52:01 2025 -0800

    rcaps: Switch prepopulated engines resolution to the new Resolver

    Switches the resolution of prepopulate template urls to use
    TemplateURLPrepopulateDataResolver instead of pulling individual
    context from the profile.

    Bug: 328040066