[CodeHealth][rust] Update `chrono-0.4.24` => `chrono-0.4.39`
This PR updates `chrono` to version `0.4.39`. This is one of the steps
to eventually allow dropping `windows-targets@0.42.2`.

Updating this crate has run into issues with the list `remove_crates` in
`gnrt_config.toml`, which had to be updated to better match the upstream
version. Addional changes were necessary to correct failures around
license file paths for some of the crates affected.

Resolves https://github.com/brave/brave-browser/issues/44035