[cr135] Fix prepulated engine function to take country id
The signature of several functions in upstream have been changed,
droppin the passing of `SearchEngineChoiceService*` as an argument, and
just taking `CountryID` directly. This has affected several tests that
have been making use of these functions. Additional, several overriden
functions had also to be corrected accordingly to reflect that.

Another aspect of this change is that `PrefService*` and `Profile*` are
now being passed around as a reference in these functions, and because
of that, several functions in brave were also update to pass these
values by ref, in order to cascade their mandatory status.

Finally, `BraveTemplateURLPrepopulateDataTest` has been updated to make
use of `search_engines::SearchEnginesTestEnvironment`, as it simplifies
the operations it was testing.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/990127eb3c219f7165b6883e914b1f2b4e340631

commit 990127eb3c219f7165b6883e914b1f2b4e340631
Author: Nicolas Dossou-Gbete <dgn@chromium.org>
Date:   Mon Feb 17 03:02:48 2025 -0800

    rcaps: Make TemplateURLPrepopulateData use rcaps for country needs

    Instead of accessing the profile country ID through
    SearchEngineChoiceService, TemplateURLPrepopulateData[::Resolver] now
    obtains it directly through RegionalCapabilitiesService. This allows
    properly updating the dependencies between these 3 services

    Bug: 328040066