[CodeHealth][rust] Update `chrono-0.4.24` => `chrono-0.4.39` (#27693)
* [CodeHealth][rust] Update `chrono-0.4.24` => `chrono-0.4.39`

This PR updates `chrono` to version `0.4.39`. This is one of the steps
to eventually allow dropping `windows-targets@0.42.2`.

Updating this crate has run into issues with the list `remove_crates` in
`gnrt_config.toml`, which had to be updated to better match the upstream
version. Addional changes were necessary to correct failures around
license file paths for some of the crates affected.

Resolves https://github.com/brave/brave-browser/issues/44035

* [CodeHealth][rust] Update `remove_crates` list with upstream

This list was not entirely matching the upstream one, which at first was
not a problem, but this has run into issues running `gnrt gen` once
`chrono` was updated. This change adds the missing `windows_*` crates
that belong to that list to the brave file.

* [CodeHealth][rust] Using a common `Apache-2.0` for some crates

For some reason some of these crates are trying to set the path for the
license file under `//brave`, rather than on the source. This change
redirects the license file to the common `Apache-2.0` license.