Fixed crash when using new region
fix https://github.com/brave/brave-browser/issues/44181

When new region is added to timezone list and it's same with user's
timezone, vpn service tries to map it to v2's name via GetMigratedNameIfNeeded().
However, new region is not included in kV1ToV2Map, it causes crash.
When it's not in that map, use that name instead.