[cr135] Spanifying `PerturbPixels`
`ImageDataBuffer` has been updated to provide `PixelData()` which is a
span. This run into failures with `PerturbPixels`.

This change spanifies the interface of `PerturbPixels`, and adjusts all
callsites to use the update interface. This has revealed the violation
of const-correctness around this methods. `ImageDataBuffer` in
particular requires a shadow file, with an override to expose the
underlying `SkPixmap` instance, to allow us to mutate the buffer.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/663ef5466d3f9741e8ada5335b3cb1857d97f631

commit 663ef5466d3f9741e8ada5335b3cb1857d97f631
Author: Fredrik Söderquist <fs@opera.com>
Date:   Mon Feb 24 17:02:44 2025 -0800

    spanification: //third_party/blink/renderer/core/html/canvas/canvas_async_blob_creator.cc

    Move the unsafe span construction into the ImageDataBuffer helper class
    instead.

    Bug: 351564777