[RELAND][CodeHealth] Use cxx20's `T::contains` pt.2 (#27786)
In `cxx20`, there is a new function in associative containers called
`contains`, which should be used as a replacement to `base::Contains`.

This reland makes use of a `base::flat_set<ContentGroup>` in one test,
as this was causing build failures on x86 builds, as the compiler would
not abide to evaluate `MakeFixedFlatSet` as `constexpr`.

Resolves https://github.com/brave/brave-browser/issues/44063