prevent flicker on node removal
This change instructs the `closing` class to not permit its visual state to reset to frame 0 of its associated keyframe animations. With this change, the `closing` class will remain in its last frame up until the point it is removed from the DOM, thus preventing a flicker at the end of its lifecycle.

Resolves https://github.com/brave/brave-browser/issues/44220