[Android] New c-tor signature for StatusBarColorController
Chromium change:
https://chromium.googlesource.com/chromium/src/+/6b88305d9871a487e87b7eb9991f5fbe36940d85

[TSLO] Fix status bar coloration in desktop windowing mode

The status bar currently always uses the default tab strip background
color on a tablet when the strip is visible irrespective of the focus
state of the activity window. The strip background color changes from
the default for an unfocused desktop window and if the status bar is
contiguous with an unfocused desktop window, it incorrectly still uses
the default color that is applied only to a focused tab strip.

This CL makes an update to ensure that the status bar uses the same
color applied to the tab strip, which can be impacted by the activity
focus state.

Before:
https://screenshot.googleplex.com/9uKWnPUYcouZFSm
After:
https://screenshot.googleplex.com/AQVoLoAkvwqVvQG

Bug: 393443576