[CodeHealth] Handle missing sources when applying patches
When running `apply_patches` and a patch fails to apply because the
source file it was created for is gone from Chromium, the tool usually
emits the following error:

```
[Error: ENOENT: no such file or directory, open '/home/redacted/dev/brave-dev/src/build/util/action_remote.py'] {
  errno: -2,
  code: 'ENOENT',
  syscall: 'open',
  path: '/home/redacted/dev/brave-dev/src/build/util/action_remote.py'
}
```

This error is not ideal as this interrupts the execution of the tool,
and it requires the patch to be deleted first, before a proper run of
`apply_patches` with error reporting can be finished.

This change adds an extra reason why patches can fail, and it flags that
reason whenever a patch is not present.

This change also corrects an existing bug where `apply_patches` adds to
the reset list a file that is already gone. That causes all the files
listed after to not be reset, which leads to patch deletions when
running `update_patches` after.

Resolves: https://github.com/brave/brave-browser/issues/44285