Update Custom Distillation Approach, Enable by Default (#27904)
* add an option to declare if an iife is desired or not
* bundle distiller scripts directly, opt-out of an iife
* provide default distillation level value
* fixes distiller execution

With distiller scripts no longer being wrapped in IIFE's (by Webpack), we have a simpler execution pattern (i.e., { %s distill() }). This means we no longer need to combine scripts, or wait for promises to resolve. Lastly, the map of host names to distillation preferences has been moved to the header, which seems to be a more appropriate location.

* sets default distillation level to FULL
* switches distiller scripts to enabled-by-default