[canary] `cr/upgrade` command-line tool (#27817)
This PR introduces a command line tool to assist with version upgrade
tasks. This initial version supports the creation of the four basic
changes necessary to get a new brave version building.

This script is capable of a) updating the version number for Chromium,
b) perform basic conflict resolution, c) create the "Update Patches"
change for the new version, d) create the "Updated strings" change.

This tool is being introduced as a starting point for making the daily
"canary" tasks easier, as well as providing to provide more robust infra
support for these tasks and for version bumps.

This tool is should be usuable both at CI and locally.

A basic run:

```
tools/cr/upgrade.py --previous=origin/master --to=135.0.7035.1
```

Changes to `.gitignore` are being introduced as this local file is
required when the tool has give up on conflict resolution and wait for
the user to resolve them.

Resolves https://github.com/brave/brave-browser/issues/44244