[canary] `--previous` defaults to upstream branch
This change removes the need to always provide a value for `--previous`
when running the upgrade tool, by defaulting its value to whatever is
the current branch's upstream branch. This is a more reliable way to get
commit messages properly generated, as it is easy to confuse the
previous branch when using terminal auto-completion.

Resolves https://github.com/brave/brave-browser/issues/44498