[canary] Complete output makeover for `upgrade.py`
This change touches on several aspects of `upgrade.py` with the purpose
of making the output for for this utility more informative of what
exactly is being done, avoiding noise.

This PR adds support for status-like output, indicating which command in
particular is being run. It also makes the commit messages the UI driver
of progress.

Resolves https://github.com/brave/brave-browser/issues/44548