[cr136] Fix `feature.cc` duplicate from `brave_stats`
At some point this failure started happening with M136

```
> gn gen /home/redacted/dev/brave-dev/src/out/redirect_cc
ERROR at //build/config/BUILDCONFIG.gn:547:5: Duplicate object file
    target(_target_type, target_name) {
    ^----------------------------------
The target //chrome/browser:browser
generates two object files with the same name:
  obj/chrome/browser/browser/features.o

It could be you accidentally have a file listed twice in the
sources. Or, depending on how your toolchain maps sources to
object files, two source files with the same name in different
directories could map to the same object file.

In the latter case, either rename one of the files or move one of
the sources to a separate source_set to avoid them both being in
the same target.
null
```

It was not possible to determine what exactly in chromium changed to
trigger this against brave's codebase, however the fix was to rename
`browser/brave_stats/features.h` to something more unique.