[canary] Adding github issue handling to 🚀 (#28077)
This change adds support for issue creation/update when running
`upgrade.py`. Using the switch `--with-github` indicates that an issue
should be created/updated by the end of the process. Additionally,
`--github-issue-only` can be used to only create/update the issue based
on the params provided.

The tool will either create a new issue, or update the link in the
existing one to show a correct googlesource link, if the one present is
not correct. For communication with GitHub services, `gh` CLI is being
used, as it is much simpler to integrate, and it provides its own
options to manage the user token. This also means that this is being
done first having users in mind, rather than to be used in the CI.

This PR does other additional fixes around the code, specially with
regards the "working version", which means the version that was in the
branch when the upgraded started.

Resolves https://github.com/brave/brave-browser/issues/44586