[canary] 🚀 Handling broken patches during `--3way` (#28128)
This PR adds corrects a failure with 🚀 in detecting that a script is
broekn and and can't work through `apply --3way`, which would cause all
patches listed after the broken patch to not be applied as well. In
general, this would result in deleted patches, revealing the issue, but
this change now introduces a new section that properly handles this.

With this new `(action needed)` section, broken patches are listed for
fixing, and when possible, the source path provided by `apply_patches`
is also provided. This is how this new section is supposed to look:

```
* Broken patches that fail to apply entirely (action needed):
    ✘ patches/build-android-gyp-dex.py.patch ➜ build/android/gyp/dex.py
```

This PR also corrects the coverage of the `--vscode` flag, to include
opening broken patches, their target files, as well as deleted patches
with vscode.

Finally, an opportunity was taken to introduce `--verbose`, which helps
during development/troubleshooting, and will allow taking a look under
the bonnet. This is coupled with new warnings that may be relevant, and
better checks around the syncing state of brave/core and chromium/src.

Resolves https://github.com/brave/brave-browser/issues/44645