removes weak pointers, adds cropping
Removes various weak pointer instances throughout the code, relying instead on pointers to the associated web contents (which are expected to live as long as our functionality needs).

This commit also introduces cropping of the final screenshot (when needed). Our "clip" dict is not honored y the DevTools Protocol when `fromSurface` is `true`. As a result, we check dimensions post-capture, and crop as needed.