[canary] 🚀 Use `dataclasses` for 3way merging
This PR introduces several data classes to better compartimentalise the
data being used in this tool. The previous code relied on a very large
method to do all the operations necessary during patch resolution.

The management of the continuation file is also being overhauled to be
its own code, separate from `PatchFailureResolver`, as the continuation
file also stores things like `working_version` which is not necessarily
part of the merge process.

**Version Type**

This type is being introduced to simplify the operations around version
in Brockit. The code is now easier to understand, and the operations are
grouped.

**Infra Mode**
This new execution mode suppresses the status page, and as a
replacement, provides terminal "keep alive" output, to allow the CI to
be kept alive.

Resolves https://github.com/brave/brave-browser/issues/44690
Resolves https://github.com/brave/brave-browser/issues/44727