[134.0.6998.118][Linux] Temporarily patch out a DCHECK.
This DCHECK causes crashes for these 2 browser tests:

     DisabledURLSanitizerTest.JSApi (../../brave/browser/url_sanitizer/url_sanitizer_browsertest.cc:239)
     EnabledURLSanitizerTest.JSApi (../../brave/browser/url_sanitizer/url_sanitizer_browsertest.cc:235)

The test crash because the clipboard data already has data of type
chromium/x-internal-source-rfh-token when another attempt to insert the
same type happens.

Patching the DCHECK out just to get the bump through, but we need to
investigate what's happening here.

Likely culprit is this upstream change:

https://chromium.googlesource.com/chromium/src/+/c7611079f28220b620f53364c0d1c4ba11539c93

commit c7611079f28220b620f53364c0d1c4ba11539c93
Author: Dominique Fauteux-Chapleau <domfc@chromium.org>
Date:   Mon Mar 17 11:02:52 2025 -0700

    Set RFH token in clipboard only after writing other data

    See crbug.com/401379144 for an explanation of the fix and fix.

    (cherry picked from commit 1f19c23b4507a98c28fee554ccaf828e82f40f04)

    Bug: 401379144, 403574896