[cr135 follow up] Fixes showing download button in toolbar.
Upstream enabled the kPinnableDownloadsButton feature that places the
download button into the pinnable toolbar container. We were hiding
this container as its content is controlled by the new toolbar customization
UI that we currently hide. Turning the flag off would be a short term
solution as both the feature flag and the standalone download toolbar
button code are being removed in cr136. Instead, don't hide the
container any more and instead limit the buttons that can be shown in
the container to the downloads button only. Also, disable context menu
on the downloads button to hide access to the customization UI.

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/34dd137e9df9e14cc9efd8dedcebac6bd6584116

commit 34dd137e9df9e14cc9efd8dedcebac6bd6584116
Author: Caroline Rising <corising@chromium.org>
Date:   Thu Feb 6 11:45:53 2025 -0800

    Make downloads pinnable to the toolbar (part 11)

    Enable the feature flag making downloads pinnable by default and fix
    tests that break as a result.

    Bug: 323962334