[canary] Add pre-run toolchains checks to 🚀
This PR adds toolchain checks to Brockit, providing alerts that
toolchain expectations may be broken for a given upgrade. With the rust
toolchain we can check with certainty whether or not we need to generate
a new toolchain tarball. With Mac, and Windows though, it is not
possible to make that determination. For this reason these checks are
not validated for continuation.

A `--ack-advisory` flag is being introduced to permit the user to
provide acknowledgment of these advisories. This is important to have
the users attention to these changes.

The advisories may look like this listing:

```
* Pre-run advisory (attention needed)
    * Windows SDK has been updated. 10.0.22621.0 ➜ 10.0.26100.0
        CL: https://chromium.googlesource.com/chromium/src/+/fca8e8720feecf9faa988b92806b574cd530b23e
        Contact DevOps regarding the new WinSDK for the hermetic toolchain. Update `env.GYP_MSVS_HASH_*` in build/commands/lib/config.js with correct hashes.
    * MacOS SDK has been updated. 15.1 ➜ 15.2
        CL: https://chromium.googlesource.com/chromium/src/+/0640527f7a4fab9947b1282565d4e8ede49b5d61
        Contact DevOps regarding the new macOS SDK for the hermetic toolchain. Update `XCODE_VERSION` and SDK version in build/mac/download_hermetic_xcode.py for the new download URL.
    * The rust toolchain has been updated.
        CL: https://chromium.googlesource.com/chromium/src/+/f9fada98083846bfbb4cffaa7483595740b93806
        Run the jobs in https://ci.brave.com/view/rust to generate a new Rust toolchain.
```

https://github.com/brave/brave-browser/issues/44887