Cleanup SidebarItem
fix https://github.com/brave/brave-browser/issues/44896

Global methods in `sidebar_item.h` could be `SidebarItem`'s method.

This cleanup is included in brave/brave-core#21081.
But it could be done in separated PR as this cleanup is not
directly related with sidebar mobile view feature.
And it could reduce original PR's size.