[cr135 follow up] Fixes duplicated Startpage search engine. (#28293)
The duplication can be seen in the search engine choice dialogs in the
settings and in the search engine NTP widget customization.

The duplication happens on upgrade from previous version (cr134) because
in the prior version Startpage engine was saved in prefs with ID we
created for it, but after the upgrade we were using the new ID that
Chromium added. That merged the new engine and the old engine into the
list.

To fix we should continue to use our old ID for Startpage instead of
switching to the Chromium's id for the engine.