[canary] Fix 🚀 `gclient sync` silent Jenkins crash
There has been some long investigation to determine why `npm run init`
was hanging when being called from Brockit. Although the proper reason
couldn't be properly ascertained, it is possible to say now that in
certain contexts, Jenkins doesn't like `gclient sync` with the
`--revision` flag.

This PR works around this issue by providing a specific flag to remove
the use of `--revision`, and manually checkout `chromium/src` to the
desired tag. The flag is named `--with_issue_44921`, in order to provide
better context of its use and rationale behind it.

Resolves https://github.com/brave/brave-browser/issues/44921