[canary] Support `--to=latest-[beta|dev|canary]` 🚀 (#28263)
This addition is a major part of getting `brockit` running daily on the
CI. This will allows to have `brockit` providing the mechanism for
polling the latest tag from `chromiumdash.appspot.com`.

The `--to` flag now has a special tag `latest` which is followed by the
release channel we want to poll the latest version from.

Resolves https://github.com/brave/brave-browser/issues/44858