Increase associated content lengths and handle content receipts (#27630)
* Increase associated content lengths

Mixtral -> 8x increase

Llama -> 8x increase

Qwen -> 8x increase

* Add trimmed tokens warning

For now this only appears after the first message and it updates over
time on that first message. So the warning could show for example that
it could only use 65% of tokens, then later change to 62% after
subsequent messages.

* Store has_trimmed_tokens in database

With DB migration to v3