[canary] Fix `--squash` being cancelled when rebasing with B🚀
The original code was creating a rebase plan based on a git log output,
and tailoring it to specific things, like setting the first line as
`reword` from `pick`, and drop regen patches. The issue with this
approach is that git has its own functionality to move commit lines
around when squashing changes.

This PR fixes this issue by providing re-entrance to B🚀 with special
flags. This allows us to us B🚀 itself the editor when calling `git` in
interactive mode, and we can edit the TODO file at that stage as we see
fit.

Resolves https://github.com/brave/brave-browser/issues/44997