[canary] Squash minor bumps with B🚀
when running a version cycle, a couple of commits accummulate along the
way, for the version bumps, and the different conflict-resolved commits.
At the end of the cycle, all these commits are squashed.

`brockit rebase` now has a flag `--squash-minor-bumps` that is used to
squash all these intermediary bumps away.

Resolves https://github.com/brave/brave-browser/issues/45147