[canary] Add a `.vpython3` wheels file for B🚀
This PR reverts the use of `DEPS` to manage python dependencies, and
rather introduces a `.vpython3` file to deploy wheels to run `brockit`
with vpython in our CI.

This PR also has an extra fix for a rare, but possible checkout failure
with the issue https://github.com/brave/brave-browser/issues/44921. This
can happen because git requires this type of notation to retain the
reference locally.

Resolves https://github.com/brave/brave-browser/issues/45120

Resolves https://github.com/brave/brave-browser/issues/45201