Set visible playlist panel's cached contents
When SidePanelWebView is created, upstream marks this content and
has assumption that it's visible if it has loaded url.
When playlist panel is closed while playing, we cache |contents_wrapper_|
to make it continue to play after closing panel.
So, it has loaded url already. Should be visible before creating
SidePanelWebView with it to avoid above CHECK failure.