[canary] Provide better defaults for better B🚀 rebase fork point (#28527)
This change corrects an issue with our current rebase operation, and its
reliance on `git merge-base fork-point`, which gets off-base when the
upstream branch for a given branch (say `cr137`, `cr138`) gets rebased
against their own upstream branch and force pushed. At that point, the
fork point returned by git is wildly off and behind where it should
actually be.

This change adds `--to-ref`/`--from-ref` as a better way to select the
range of changes that we want to rebase. With `--to-ref` it defaults to
`@upstream` when nothing is provided, and it represents the branch we
are rebasing the version on top.

With `--from-ref` we default to `@previous-major` when dealing with
major lifts, or just to `@previous` when dealing with minor bumps.
`@previous-major` is being introduced to find exactly the first change
that changed the major version.

This PR also removes the "dry-run" attempt in determining if there could
be conflicts when rebasing a branch as that warning was entirely
off base.

Resolves https://github.com/brave/brave-browser/issues/45226