[canary] Add `Version` tests to B🚀 (#28531)
This PR adds tests for `Version` to `brockit`. For convenience,
`Version` has been moved into its own module, and also a few other bits
and pieces that were associated with it.

This change also touches a lot on `Repository`, because these two
modules make heavy use of a common test machinery for the two. This
means adjusting both tests to make better reuse of code.

Finally, `Repository.chromium()`/`Repository.brave()` are being changed
into globals inside `repository`. This is to avoid instatiations on each
use, and to follow the advise in
https://google.github.io/styleguide/pyguide.html#2174-decision regarding
the use of `@classmethod`

Resolves https://github.com/brave/brave-browser/issues/45231