[canary] Add `--tag` arg to `npm run build_rust_toolchain_aux` (#28530)
The addition of this argument will allow us to checkout a different
Chromium version, as a tantive version. This is useful when generating
the toolchain for a version bump that the branch itself is not fully
up-to-date with, but it is close enough to provide hermetic toolchains
necessary to run the toolchain job with.

Resolves https://github.com/brave/brave-browser/issues/45237