[canary] Adding some `Patchfile` test to B🚀
This change adds a few tests to cover the code in `Patchfile`. This
doesn't cover all cases to these files, but it is a good start, and it
manages to cover the core operation of applying patches.

A lot of test machinery is being introduced to provide some fake support
ot what normally would be done by checkout framework managed in node.

Resolves https://github.com/brave/brave-browser/issues/45232