[canary][CodeHealth] Tooling for posting `PRESUBMIT` warnings to PRs
This is a very basic tooling that will be hosted in `brave-core` to
allow us to process the report generated by a `PRESUBMIT` run, and post
it as comments in PR.

This tool has a basic hashing scheme to make sure we don't post the same
comment every time this tool is called. We also delete comments once
they are not being alerted for anymore. This for now is the simplest way
to make sure that new alerts do cause a PR email notification, but at
the same time it removes old ones, so reviewers don't have to open the
pipeline to double check which alerts are still active.

Resolves https://github.com/brave/brave-browser/issues/45402