[canary] Reduce `--no-conflict-change` to a validation
This flag was first introduced as a driver of how the `brockit` would
behave when doing continuations. However it simplifies a lot to reduce
this flag to be merely a validation for when there's nothing to commit
as a conflict-resolved change.

A further advantage of having this flag as merely a validation as users
could in fact provide the flag wrongly (for instance, terminal
completion), and this would change the way the process would flow. With
this change the process as a whole stays the same, and the flag merely
makes it more permissive.

Resolves https://github.com/brave/brave-browser/issues/45317