[rust][cr137] Use upstream `hashbrown` and bump version
This change drops the vendored copy of `hashbrown` in `brave-core` and
makes use of the upstream one. Unfortunately the upstream one does not
use many features from this crate, some of which require extra deps.
To correct that, this change adds a patch to the upstream `BUILD.gn`.

Chromium changes:
https://chromium.googlesource.com/chromium/src/+/05c5c4f485b52f12f35f7a603b6505fbdd363a4d

commit 05c5c4f485b52f12f35f7a603b6505fbdd363a4d
Author: amarjotgill <amarjotgill@google.com>
Date:   Fri Feb 14 08:36:21 2025 -0800

    Add Sfv version 0.10.4 third party crate

    Bug: 391378116
    Change-Id: I5eed95c02e8c05cb2fd87855388ba7c55efbf1a6
    Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6265583
    Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
    Commit-Queue: Amarjot Gill <amarjotgill@chromium.org>
    Cr-Commit-Position: refs/heads/main@{#1420510}