[canary] Use exceptions to communicate B🚀 failures
This PR changes the use of bool to signal success/failure when running
`brockit` tasks, and replaces them with exceptions. The main issue with
the use of return values is that it was not possible to use return
values in every case. For instance, some failures were due to
validations in the task constructor, that belong in there. This change
will make testing these tasks a bit easier too.

Additionally, in this PR we include a fix to `Repository.run_git` where
`no_trim` was not being passed to terminal when the repository was
brave.

Resolves https://github.com/brave/brave-browser/issues/45596